dysfunctional state -Wall

+++ testing/tests/moduletests/BAS/BAS_01.cpp
@@ -74,7 +74,7 @@ namespace test
         L_CHECK_EQUAL( 0, a.size())
 
         STEP(2)
-        for (INT32 i = 0; i < a.capacity(); ++i)
+        for (UINT16 i = 0; i < a.capacity(); ++i)
         {
             L_CHECK_TRUE(a.hasSpace())
             const size_t p = a.addT<CData>(-i, i);
@@ -86,7 +86,7 @@ namespace test
         STEP(3)
         //  test unsorted data as loaded
         SUBSTEPS()
-        for (INT32 i = 0; i < a.size(); ++i)
+        for (UINT16 i = 0; i < a.size(); ++i)
         {
             LSTEP(i)
             const IData& d = a.at(i);
@@ -101,7 +101,7 @@ namespace test
         a.sort();
         const INT32 of = a.size() - 1;
         SUBSTEPS()
-        for (INT32 i = 0; i < a.size(); ++i)
+        for (UINT16 i = 0; i < a.size(); ++i)
         {
             LSTEP(i)
             const IData& d = a.at(i);
@@ -114,7 +114,7 @@ namespace test
         //  apply find() to const object
         const IDataArray& c = a;
         SUBSTEPS()
-        for (INT32 i = 0; i < c.size(); ++i)
+        for (UINT16 i = 0; i < c.size(); ++i)
         {
             LSTEP(i)
             const PosRes p = c.find(c.at(i));
@@ -163,7 +163,7 @@ namespace test
 
         STEP(2)
         //  load data
-        for (INT32 i = 0; i < a.capacity(); ++i)
+        for (UINT16 i = 0; i < a.capacity(); ++i)
         {
             a.addT<CData>(-i, i);
         }
@@ -176,7 +176,7 @@ namespace test
         L_CHECK_TRUE(ok)
 
         SUBSTEPS()
-        for (INT32 i = 0; i < a.size(); ++i)
+        for (UINT16 i = 0; i < a.size(); ++i)
         {
             LSTEP(i)
             const IData& d = a.at(i);
@@ -215,7 +215,7 @@ namespace test
         
         CDataArray a;
         STEP(1)
-        for (INT32 i = 0; i < a.capacity(); ++i)
+        for (UINT16 i = 0; i < a.capacity(); ++i)
         {
             a.add(-i, i);
         }
